compile fixes
svn path=/trunk/kdelibs/; revision=257554

+++ khtml/rendering/render_block.h
@@ -135,7 +135,7 @@ public:
 
     virtual unsigned short lineWidth(int y) const;
     virtual int lowestPosition() const;
-    virtual int rightmostPosition() const;
+    virtual short rightmostPosition() const;
 
     int rightOffset() const;
     int rightRelOffset(int y, int fixedOffset, int *heightRemaining = 0) const;
