+++ khtml/rendering/render_text.h
@@ -230,7 +230,7 @@ public:
     virtual short marginLeft() const { return style()->marginLeft().minWidth(0); }
     virtual short marginRight() const { return style()->marginRight().minWidth(0); }
 
-    virtual int rightmostPosition() const;
+    virtual short rightmostPosition() const;
 
     virtual void repaint(bool immediate=false);
